fix(ci): cast TRIAL_DAYS to int in config and fix PHPStan PasswordBroker type [v-ci-fix]
- config/plans.php: cast env('TRIAL_DAYS') to int so config always
  returns int regardless of .env string parsing — fixes PlanConfigTest
- AdminUsersController: add @var PasswordBroker annotation for
  Password::broker() to resolve PHPStan interface/concrete mismatch

+++ config/plans.php
@@ -210,7 +210,7 @@ return [
     */
     'trial' => [
         'enabled' => env('TRIAL_ENABLED', true),
-        'days' => env('TRIAL_DAYS', 14),
+        'days' => (int) env('TRIAL_DAYS', 14),
         'tier' => 'pro', // which tier to grant during trial
     ],
 ];
